daily_screener/requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pytrends>=4.9.2
